Move bal/ pages components to components directory
